# public/openapi.json
{"openapi":"3.0.0","info":{"title":"iDoc API Reference","version":"","description":"iDoc Api secification and documentation.","termsOfService":"","license":null,"contact":[],"x-logo":{"url":"https:\/\/avatars.githubusercontent.com\/u\/26082509?v=4","altText":"iDoc API Reference","backgroundColor":""}},"components":{"securitySchemes":{"BearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}},"schemas":[]},"servers":[{"url":"http:\/\/localhost:8000","description":"Documentation generator server."},{"url":"http:\/\/test.example.com","description":"Test server."}],"paths":{"\/api\/login":{"post":{"security":[{"BearerAuth":[]}],"tags":["general"],"operationId":"Login","description":"","requestBody":{"description":"","required":true,"content":{"application\/json":{"schema":{"type":"object","required":["username","password"],"properties":{"username":{"type":"string","example":"\"engineer@mail.com\"","description":"The email of the user."},"password":{"type":"string","example":"\"password\"","description":"The name of the role."}},"example":{"username":"\"engineer@mail.com\"","password":"\"password\""}}}}},"parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"status":"success","success":true,"message":"Login success!","data":null}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X POST \"http:\/\/localhost:8000\/api\/login\" \\\n    -H \"Authorization: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\" \\\n\n    -d \"username\"=\"\"engineer@mail.com\"\" \\\n    -d \"password\"=\"\"password\"\" "},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/login\");\n\nlet headers = {\n    \"Authorization\": \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nlet body = JSON.stringify({\n    \"username\": \"&quot;engineer@mail.com&quot;\",\n    \"password\": \"&quot;password&quot;\",\n})\n\nfetch(url, {\n    method: \"POST\",\n    headers: headers,\n    body: body\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Authorization\" => \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$body = [\n    \"username\" =>  \"&quot;engineer@mail.com&quot;\",\n    \"password\" =>  \"&quot;password&quot;\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/login\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"POST\");\ncurl_setopt($ch, CURLOPT_POSTFIELDS, $body);\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/register":{"post":{"security":[{"BearerAuth":[]}],"tags":["general"],"operationId":"Register","description":"","requestBody":{"description":"","required":true,"content":{"application\/json":{"schema":{"type":"object","required":["role","name","email","password","confirm_password"],"properties":{"role":{"type":"number","example":1,"description":"The role of the user."},"name":{"type":"string","example":"\"Ahmad Farid Imam Zamani\"","description":"The name of the user."},"email":{"type":"string","example":"\"engineer@mail.com\"","description":"The email of the user."},"password":{"type":"string","example":"\"password\"","description":"The password of the user."},"confirm_password":{"type":"string","example":"\"password\"","description":"The password confirmation of the role."}},"example":{"role":1,"name":"\"Ahmad Farid Imam Zamani\"","email":"\"engineer@mail.com\"","password":"\"password\"","confirm_password":"\"password\""}}}}},"parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"status":"success","success":true,"message":"Register success!","data":null}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X POST \"http:\/\/localhost:8000\/api\/register\" \\\n    -H \"Authorization: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\" \\\n\n    -d \"role\"=\"1\" \\\n    -d \"name\"=\"\"Ahmad Farid Imam Zamani\"\" \\\n    -d \"email\"=\"\"engineer@mail.com\"\" \\\n    -d \"password\"=\"\"password\"\" \\\n    -d \"confirm_password\"=\"\"password\"\" "},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/register\");\n\nlet headers = {\n    \"Authorization\": \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nlet body = JSON.stringify({\n    \"role\": \"1\",\n    \"name\": \"&quot;Ahmad Farid Imam Zamani&quot;\",\n    \"email\": \"&quot;engineer@mail.com&quot;\",\n    \"password\": \"&quot;password&quot;\",\n    \"confirm_password\": \"&quot;password&quot;\",\n})\n\nfetch(url, {\n    method: \"POST\",\n    headers: headers,\n    body: body\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Authorization\" => \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$body = [\n    \"role\" =>  \"1\",\n    \"name\" =>  \"&quot;Ahmad Farid Imam Zamani&quot;\",\n    \"email\" =>  \"&quot;engineer@mail.com&quot;\",\n    \"password\" =>  \"&quot;password&quot;\",\n    \"confirm_password\" =>  \"&quot;password&quot;\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/register\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"POST\");\ncurl_setopt($ch, CURLOPT_POSTFIELDS, $body);\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/role":{"get":{"security":[{"BearerAuth":[]}],"tags":["general"],"operationId":"Get Roles","description":"get roles data","parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"status":"success","success":true,"message":"get role data success","data":[{"id":1,"name":"Software Engineer","created_at":null,"updated_at":null},{"id":2,"name":"HR Manager","created_at":null,"updated_at":null},{"id":3,"name":"HR Staff","created_at":null,"updated_at":"2023-12-18T08:04:41.000000Z"}]}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X GET -G \"http:\/\/localhost:8000\/api\/role\" \\\n    -H \"Authorization: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\"\n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/role\");\n\nlet headers = {\n    \"Authorization\": \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"GET\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Authorization\" => \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/role\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]},"post":{"security":[{"BearerAuth":[]}],"tags":["general"],"operationId":"Create Role","description":"Add a new item on the role.","requestBody":{"description":"Add a new item on the role.","required":true,"content":{"application\/json":{"schema":{"type":"object","required":["name"],"properties":{"name":{"type":"string","example":"\"Software Engineer\"","description":"The name of the role."}},"example":{"name":"\"Software Engineer\""}}}}},"parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"status":"success","success":true,"message":"role added successfully!","data":{"name":"Software Engineer","updated_at":"2023-12-18T09:29:14.000000Z","created_at":"2023-12-18T09:29:14.000000Z","id":6}}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X POST \"http:\/\/localhost:8000\/api\/role\" \\\n    -H \"Authorization: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\" \\\n\n    -d \"name\"=\"\"Software Engineer\"\" "},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/role\");\n\nlet headers = {\n    \"Authorization\": \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nlet body = JSON.stringify({\n    \"name\": \"&quot;Software Engineer&quot;\",\n})\n\nfetch(url, {\n    method: \"POST\",\n    headers: headers,\n    body: body\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Authorization\" => \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$body = [\n    \"name\" =>  \"&quot;Software Engineer&quot;\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/role\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"POST\");\ncurl_setopt($ch, CURLOPT_POSTFIELDS, $body);\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/role\/{role}":{"put":{"security":[{"BearerAuth":[]}],"tags":["general"],"operationId":"Update Role","description":"Update item on the employee.","requestBody":{"description":"Update item on the employee.","required":true,"content":{"application\/json":{"schema":{"type":"object","required":["name"],"properties":{"name":{"type":"string","example":"\"Software Engineer\"","description":"The name of the role."}},"example":{"name":"\"Software Engineer\""}}}}},"parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"status":"success","success":true,"message":"Success to update role","data":null}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X PUT \"http:\/\/localhost:8000\/api\/role\/{role}\" \\\n    -H \"Authorization: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\" \\\n\n    -d \"name\"=\"\"Software Engineer\"\" "},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/role\/{role}\");\n\nlet headers = {\n    \"Authorization\": \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nlet body = JSON.stringify({\n    \"name\": \"&quot;Software Engineer&quot;\",\n})\n\nfetch(url, {\n    method: \"PUT\",\n    headers: headers,\n    body: body\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Authorization\" => \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$body = [\n    \"name\" =>  \"&quot;Software Engineer&quot;\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/role\/{role}\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"PUT\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]},"delete":{"security":[{"BearerAuth":[]}],"tags":["general"],"operationId":"Delete Role","description":"Delete item on the role.","parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"status":"success","success":true,"message":"Success to delete role","data":null}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X DELETE \"http:\/\/localhost:8000\/api\/role\/{role}\" \\\n    -H \"Authorization: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\"\n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/role\/{role}\");\n\nlet headers = {\n    \"Authorization\": \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"DELETE\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Authorization\" => \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/role\/{role}\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"DELETE\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/employee":{"get":{"security":[{"BearerAuth":[]}],"tags":["general"],"operationId":"Get Employee","description":"get employees data","parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"status":"success","success":true,"message":"get employees data success","data":[{"id":29,"role":1,"name":"Farid Zamani","salary":"15000000","created_at":"2023-12-18T09:00:32.000000Z","updated_at":"2023-12-18T09:00:32.000000Z"},{"id":1,"role":1,"name":"Farid Zamani","salary":"15000000","created_at":"2023-12-16T10:07:39.000000Z","updated_at":"2023-12-18T06:28:43.000000Z"}]}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X GET -G \"http:\/\/localhost:8000\/api\/employee\" \\\n    -H \"Authorization: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\"\n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/employee\");\n\nlet headers = {\n    \"Authorization\": \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"GET\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Authorization\" => \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/employee\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]},"post":{"security":[{"BearerAuth":[]}],"tags":["general"],"operationId":"Create Employee","description":"Add a new item to the employee.","requestBody":{"description":"Add a new item to the employee.","required":true,"content":{"application\/json":{"schema":{"type":"object","required":["role","name","salary"],"properties":{"role":{"type":"number","example":1,"description":"The role of the employee."},"name":{"type":"string","example":"Farid Zamani","description":"The name of the employee."},"salary":{"type":"number","example":15000000,"description":"The salary of the employee."}},"example":{"role":1,"name":"Farid Zamani","salary":15000000}}}}},"parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"status":"success","success":true,"message":"employee added successfully!","data":{"role":1,"name":"Farid Zamani","salary":15000000,"updated_at":"2023-12-18T09:00:32.000000Z","created_at":"2023-12-18T09:00:32.000000Z","id":1}}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X POST \"http:\/\/localhost:8000\/api\/employee\" \\\n    -H \"Authorization: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\" \\\n\n    -d \"role\"=\"1\" \\\n    -d \"name\"=\"Farid Zamani\" \\\n    -d \"salary\"=\"15000000\" "},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/employee\");\n\nlet headers = {\n    \"Authorization\": \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nlet body = JSON.stringify({\n    \"role\": \"1\",\n    \"name\": \"Farid Zamani\",\n    \"salary\": \"15000000\",\n})\n\nfetch(url, {\n    method: \"POST\",\n    headers: headers,\n    body: body\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Authorization\" => \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$body = [\n    \"role\" =>  \"1\",\n    \"name\" =>  \"Farid Zamani\",\n    \"salary\" =>  \"15000000\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/employee\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"POST\");\ncurl_setopt($ch, CURLOPT_POSTFIELDS, $body);\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/employee\/{employee}":{"put":{"security":[{"BearerAuth":[]}],"tags":["general"],"operationId":"Update Employee","description":"Update item on the employee.","requestBody":{"description":"Update item on the employee.","required":true,"content":{"application\/json":{"schema":{"type":"object","required":["role","name","salary"],"properties":{"role":{"type":"number","example":1,"description":"The role of the employee."},"name":{"type":"string","example":"Farid Zamani","description":"The name of the employee."},"salary":{"type":"number","example":15000000,"description":"The salary of the employee."}},"example":{"role":1,"name":"Farid Zamani","salary":15000000}}}}},"parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"status":"success","success":true,"message":"employee updated successfully!","data":null}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X PUT \"http:\/\/localhost:8000\/api\/employee\/{employee}\" \\\n    -H \"Authorization: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\" \\\n\n    -d \"role\"=\"1\" \\\n    -d \"name\"=\"Farid Zamani\" \\\n    -d \"salary\"=\"15000000\" "},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/employee\/{employee}\");\n\nlet headers = {\n    \"Authorization\": \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nlet body = JSON.stringify({\n    \"role\": \"1\",\n    \"name\": \"Farid Zamani\",\n    \"salary\": \"15000000\",\n})\n\nfetch(url, {\n    method: \"PUT\",\n    headers: headers,\n    body: body\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Authorization\" => \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$body = [\n    \"role\" =>  \"1\",\n    \"name\" =>  \"Farid Zamani\",\n    \"salary\" =>  \"15000000\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/employee\/{employee}\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"PUT\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]},"delete":{"security":[{"BearerAuth":[]}],"tags":["general"],"operationId":"Delete Employee","description":"Delete item on the employee.","parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"status":"success","success":true,"message":"Success to delete employee","data":null}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X DELETE \"http:\/\/localhost:8000\/api\/employee\/{employee}\" \\\n    -H \"Authorization: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\"\n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/employee\/{employee}\");\n\nlet headers = {\n    \"Authorization\": \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"DELETE\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Authorization\" => \"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJod\",\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/employee\/{employee}\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"DELETE\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}}},"x-tagGroups":[]}